+++ misosoup/library/analysis.py
@@ -65,7 +65,41 @@ def find_focal_strain_growth(data_frame: pd.DataFrame):
     return data_frame
 
 
-def count_viable_environments(data_frame: pd.DataFrame):
+def get_unique_strains(data_frame: pd.DataFrame):
+    """Get all strains included in the dataset.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    regex = re.compile(r"Growth_(\w+)")
+    return [regex.match(c).groups()[0] for c in data_frame.columns if regex.match(c)]
+
+
+def get_unique_focal_strains(data_frame: pd.DataFrame):
+    """Get all focal strains included in the dataset.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    return data_frame.reset_index().strain.unique()
+
+
+def get_unique_carbon_sources(data_frame: pd.DataFrame):
+    """Get all unique carbon sources in the dataset.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    return data_frame.reset_index().carbon_source.unique()
+
+
+def count_carbon_sources_for_strain(data_frame: pd.DataFrame):
     """Count the number of viable environments for each strain.
 
     Parameters
@@ -73,12 +107,134 @@ def count_viable_environments(data_frame: pd.DataFrame):
     data_frame : pandas.DataFrame
         A solution dataframe from misosoup.
     """
+    df = data_frame.reset_index()
+    strains = get_unique_strains(df)
+    unique_carbon_sources = [
+        len(df[df[f"Growth_{strain}"] > 0].carbon_source.unique()) for strain in strains
+    ]
+    return pd.DataFrame({"strain": strains, "carbon_sources": unique_carbon_sources})
+
+
+def count_carbon_sources_as_focal_strain(data_frame: pd.DataFrame):
+    """Count the number of viable environments as focal strain.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    df = data_frame.reset_index()
+    strains = get_unique_strains(df)
+    unique_carbon_sources = [
+        len(df[(df.strain == strain) & (df.growth_rate > 0)].carbon_source.unique())
+        for strain in strains
+    ]
+    return pd.DataFrame(
+        {"strain": strains, "focal_carbon_sources": unique_carbon_sources}
+    )
+
+
+def count_carbon_sources_as_supplier_strain(data_frame: pd.DataFrame):
+    """Count the number of viable environments as supplier strain.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    df = data_frame.reset_index()
+    strains = get_unique_strains(df)
+    unique_carbon_sources = [
+        len(
+            df[
+                (df.strain != strain) & (df[f"Growth_{strain}"] > 0)
+            ].carbon_source.unique()
+        )
+        for strain in strains
+    ]
+    return pd.DataFrame(
+        {"strain": strains, "supplier_carbon_sources": unique_carbon_sources}
+    )
+
+
+def count_communities_as_supplier_strain(data_frame: pd.DataFrame):
+    """Count the number of communities as supplier strain.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    df = data_frame.reset_index()
+    strains = get_unique_strains(df)
+    unique_communities = [
+        len(df[(df.strain != strain) & (df[f"Growth_{strain}"] > 0)])
+        for strain in strains
+    ]
+    return pd.DataFrame({"strain": strains, "communities": unique_communities})
+
+
+def count_isolated_strains_on_carbon_source(data_frame: pd.DataFrame):
+    """Count the number of strains in isolation on a carbon source.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    df = data_frame.reset_index()
+    indicator_columns = [column for column in df.columns if column.startswith("y_")]
+    return pd.DataFrame(
+        df[df[indicator_columns].sum(axis=1) == 1]
+        .groupby("carbon_source")
+        .size()
+        .rename("isolated_strains")
+    )
+
+
+def count_non_redundant_supplying_communities_on_carbon_source(
+    data_frame: pd.DataFrame,
+):
+    """Count the number of non-redundant supplying communities on a carbon source.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    df = data_frame.reset_index()
+    supplying_communities = pd.DataFrame(
+        {
+            "carbon_source": df.carbon_source,
+            "supplying_community": get_suppliers(df),
+        }
+    )
     return (
-        data_frame[data_frame.growth_rate > 1e-4]
-        .groupby(["carbon_source", "strain"])
-        .size()
-        .groupby("strain")
-        .size()
+        supplying_communities.groupby("carbon_source")
+        .apply(lambda group: len(frozenset(group.supplying_communities)))
+        .rename("supplying_communities")
+    )
+
+
+def count_exchanged_metabolites_on_carbon_source(data_frame: pd.DataFrame):
+    """Count the number of metabolites exchanged on a carbon source.
+
+    Parameters
+    ----------
+    data_frame : pandas.DataFrame
+        A solution dataframe from misosoup.
+    """
+    df = data_frame.reset_index()
+    metabolites_exchanged = pd.DataFrame(
+        {
+            "carbon_source": df.carbon_source,
+            "metabolites_exchanged": compute_crossfeed(df),
+        }
+    )
+    return (
+        metabolites_exchanged.groupby("carbon_source")
+        .apply(lambda group: len(frozenset(group.metabolites_exchanged)))
+        .rename("metabolites_exchanged")
     )