+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Report feasible communities, even if optimization fails
 - Add function to get `supplier` and `community` frozensets.
 - Update crossfeed to report as frozensets.
+- Add functions to gather statistics.
 
 ## 1.2.0 --- Caching Interference (Mar 1, 2022)